Correct `sumGPUs` to include MIGs in count (#3933)
